Move `@Builder` annotation to `annotations` package.
