+++ processor/src/test/resources/record/nested/inputs/Organisation.java
@@ -1,6 +1,6 @@
 package test.example;
 
-import io.bothy.tog.Builder;
+import io.bothy.tog.annotations.Builder;
 import test.example.OrganisationDepartmentPersonBuilder;
 
 public class Organisation {
